Fix styles after testing with multiple tabs in Widget.
